Refactor significantly to facilitate simple file dumps.
This makes it easier to dump arrays to file and then load separately in
Python. While the built in plots are useful, sometimes it's better to be
able to play directly with data. This streamlines the process of getting
an array into a file formatted arbitrarily.

In particular, this adds a numpy format, so that the data are available
with a single python command.